node: move Node to its own package
